pyproject.toml
[tool.poetry]
name = "mermaid-electron"
version = "0.0.0"
description = ""
authors = ["whinee"]

[tool.poetry.dependencies]
python = ">=3.10,<4"

rich = "^13.3.3"
pillow = "^9.5.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]

[tool.mypy]
disable_error_code = ["import", "attr-defined", "no-untyped-call"]

[tool.ruff]
select = [
    "ANN",
    "B",
    "BLE",
    "C4",
    "C90",
    "COM",
    "D",
    # "DTZ",
    "E",
    "F",
    "I",
    "INP",
    "N",
    "PIE",
    "Q",
    "RET",
    "RSE",
    "RUF",
    "S",
    "UP"
]

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = [
    "ANN",
    "B",
    "C4",
    "COM",
    "D",
    "E",
    "F",
    "I",
    "PIE",
    "Q",
    "RET",
    "RSE",
    "RUF",
    "UP"
]
unfixable = []

ignore = [
    "ANN001",
    "ANN002",
    "ANN003",
    "ANN101",
    "ANN201",
    "ANN202",
    "ANN401",
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    "D202",
    "D203",
    "D212",
    "D401",
    "D406",
    "D407",
    "D417",
    "E501",
    "I002",
    "N812",
    "N818",
    "RET503",
    "S101",
    "UP007",
    "UP032",
]

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "pyenv",
    "venv",
]

line-length = 88

target-version = "py310"

[tool.ruff.mccabe]
max-complexity = 5

[tool.ruff.pycodestyle]
ignore-overlong-task-comments = false